Style customer view header
