+++ src/css/base.scss
@@ -343,6 +343,37 @@ header {
   font-size: 14px;
 }
 
+// CUSTOMER View
+
+.header-info {
+  display: flex;
+  flex-direction: column;
+}
+
+.customer-header {
+  display: flex;
+  justify-content: flex-end;
+  font-family: 'Comfortaa', cursive;
+  border-bottom: solid black 1px;
+}
+
+.customer-status {
+  display: flex;
+  align-items: center;
+  margin-left: 5%;
+  margin-right: 6%;
+  margin-top: 1%;
+}
+
+.customer-welcome {
+  font-size: 22px;
+  margin-bottom: 2%;
+}
+
+.gold {
+  font-size: 11px;
+}
+
 .hidden {
   display: none;
 }
